streamlit/requirements.txt
snowflake-connector-python
streamlit
pandas
plotly
posit-sdk
